+++ src/Audit.EntityFramework/ConfigurationApi/AuditEntityMapping.cs
@@ -7,7 +7,9 @@ namespace Audit.EntityFramework.ConfigurationApi
     public class AuditEntityMapping : IAuditEntityMapping
     {
         private Dictionary<Type, MappingInfo> _mapping = new Dictionary<Type, MappingInfo>();
+        private List<KeyValuePair<Func<EventEntry, bool>, MappingInfo>> _explicitMapping = new List<KeyValuePair<Func<EventEntry, bool>, MappingInfo>>();
         private Func<AuditEvent, EventEntry, object, bool> _commonAction = null;
+        
 
         public IAuditEntityMapping Map<TSourceEntity, TAuditEntity>()
         {
@@ -96,6 +98,34 @@ namespace Audit.EntityFramework.ConfigurationApi
             return this;
         }
 
+        public IAuditEntityMapping MapExplicit<TAuditEntity>(Func<EventEntry, bool> predicate, Action<EventEntry, TAuditEntity> entityAction = null)
+        {
+            _explicitMapping.Add(new KeyValuePair<Func<EventEntry, bool>, MappingInfo>(predicate, new MappingInfo()
+            {
+                TargetType = typeof(TAuditEntity),
+                Action = (ev, entry, entity) =>
+                {
+                    entityAction?.Invoke(entry, (TAuditEntity)entity);
+                    return true;
+                }
+            }));
+            return this;
+        }
+
+        public IAuditEntityMapping MapTable<TAuditEntity>(string tableName, Action<EventEntry, TAuditEntity> entityAction = null)
+        {
+            _explicitMapping.Add(new KeyValuePair<Func<EventEntry, bool>, MappingInfo>(ent => ent.Table?.Equals(tableName) == true, new MappingInfo()
+            {
+                TargetType = typeof(TAuditEntity),
+                Action = (ev, entry, entity) =>
+                {
+                    entityAction?.Invoke(entry, (TAuditEntity)entity);
+                    return true;
+                }
+            }));
+            return this;
+        }
+
         public void AuditEntityAction(Action<AuditEvent, EventEntry, object> entityAction)
         {
             _commonAction = (ev, ent, auditEntity) =>
@@ -133,18 +163,55 @@ namespace Audit.EntityFramework.ConfigurationApi
             };
         }
 
+        internal Func<EventEntry, Type> GetExplicitMapper()
+        {
+            return (e) =>
+            {
+                if (_explicitMapping == null)
+                {
+                    return null;
+                }
+                foreach(var em in _explicitMapping)
+                {
+                    if (em.Key.Invoke(e))
+                    {
+                        return em.Value.TargetTypeMapper?.Invoke(e);
+                    }
+                }
+                return null;
+            };
+        }
+
+        // Returns a generic action that executes the specific action for the mapping and the common action
         internal Func<AuditEvent, EventEntry, object, bool> GetAction()
         {
             return (ev, ent, auditEntity) =>
             {
                 MappingInfo map = null;
                 bool include = true;
-                var entityType = ent.EntityType;
-                if (entityType != null && _mapping.TryGetValue(entityType, out map))
+                bool mappedExplicitly = false;
+                if (_explicitMapping != null)
                 {
-                    if (map.Action != null)
+                    foreach (var em in _explicitMapping)
                     {
-                        include = map.Action.Invoke(ev, ent, auditEntity);
+                        if (em.Key.Invoke(ent))
+                        {
+                            map = em.Value;
+                            include = map.Action.Invoke(ev, ent, auditEntity);
+                            mappedExplicitly = true;
+                            break;
+                        }
+                    }
+                }
+                if (!mappedExplicitly)
+                {
+                    var entityType = ent.EntityType;
+                    if (entityType != null && _mapping.TryGetValue(entityType, out map))
+                    {
+                        if (map.Action != null)
+                        {
+                            include = map.Action.Invoke(ev, ent, auditEntity);
+                        }
                     }
                 }
                 if (include && _commonAction != null)