+++ test/Audit.EntityFramework.UnitTest/Tag.cs
@@ -12,18 +12,18 @@ namespace Audit.EntityFramework.UnitTest
     using System;
     using System.Collections.Generic;
     
-    public partial class Post
+    public partial class Tag
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
-        public Post()
+        public Tag()
         {
-            this.Tags = new HashSet<Tag>();
+            this.Posts = new HashSet<Post>();
         }
     
-        public int PostId { get; set; }
-        public string Text { get; set; }
+        public int TagId { get; set; }
+        public string Name { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
-        public virtual ICollection<Tag> Tags { get; set; }
+        public virtual ICollection<Post> Posts { get; set; }
     }
 }